Milestone 2 Completada

+++ src/com/rockets/controller/AccelerateMultiThread.java
@@ -6,10 +6,10 @@ public class AccelerateMultiThread implements Runnable {
 	
 	private String id;
 	private int booster;
-	private int power;
+	private double power;
 
 
-	public AccelerateMultiThread(String id,int booster, int power) {
+	public AccelerateMultiThread(String id,int booster, double power) {
 		
 		this.id=id;
 		this.booster=booster;
